render: Move filter size calculation from wgpu to render/swf
