+++ render/wgpu/src/filters/drop_shadow.rs
@@ -4,7 +4,7 @@ use crate::filters::blur::BlurFilter;
 use crate::filters::glow::GlowFilter;
 use crate::filters::FilterSource;
 use crate::surface::target::CommandTarget;
-use swf::{DropShadowFilter as DropShadowFilterArgs, Rectangle};
+use swf::DropShadowFilter as DropShadowFilterArgs;
 
 /// Drop shadow is just Glow with an offset.
 /// None of this strictly needs to be a struct,
@@ -12,31 +12,6 @@ use swf::{DropShadowFilter as DropShadowFilterArgs, Rectangle};
 pub struct DropShadowFilter;
 
 impl DropShadowFilter {
-    pub fn calculate_dest_rect(
-        filter: &DropShadowFilterArgs,
-        source_rect: Rectangle<i32>,
-        blur_filter: &BlurFilter,
-        glow_filter: &GlowFilter,
-    ) -> Rectangle<i32> {
-        let mut result =
-            glow_filter.calculate_dest_rect(&filter.inner_glow_filter(), source_rect, blur_filter);
-        let distance = filter.distance.to_f32();
-        let angle = filter.angle.to_f32();
-        let x = (angle.cos() * distance).ceil() as i32;
-        let y = (angle.sin() * distance).ceil() as i32;
-        if x < 0 {
-            result.x_min += x;
-        } else {
-            result.x_max += x;
-        }
-        if y < 0 {
-            result.y_min += y;
-        } else {
-            result.y_max += y;
-        }
-        result
-    }
-
     pub fn apply(
         descriptors: &Descriptors,
         texture_pool: &mut TexturePool,
