+++ swf/src/types/drop_shadow_filter.rs
@@ -1,4 +1,6 @@
-use crate::{BlurFilter, BlurFilterFlags, Color, Fixed16, Fixed8, GlowFilter, GlowFilterFlags};
+use crate::{
+    BlurFilter, BlurFilterFlags, Color, Fixed16, Fixed8, GlowFilter, GlowFilterFlags, Rectangle,
+};
 use bitflags::bitflags;
 
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -59,6 +61,25 @@ impl DropShadowFilter {
             strength: self.strength,
             flags,
         }
+    }
+
+    pub fn calculate_dest_rect(&self, source_rect: Rectangle<i32>) -> Rectangle<i32> {
+        let mut result = self.inner_glow_filter().calculate_dest_rect(source_rect);
+        let distance = self.distance.to_f32();
+        let angle = self.angle.to_f32();
+        let x = (angle.cos() * distance).ceil() as i32;
+        let y = (angle.sin() * distance).ceil() as i32;
+        if x < 0 {
+            result.x_min += x;
+        } else {
+            result.x_max += x;
+        }
+        if y < 0 {
+            result.y_min += y;
+        } else {
+            result.y_max += y;
+        }
+        result
     }
 }
